requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
Django==4.2.26
django-allauth==65.13.1
django-countries==8.2.0
django-environ==0.12.0
django-model-utils==5.0.0
django-modeltranslation==0.19.17
django-paypal==2.1
django-robots==6.1
exceptiongroup==1.3.1
fonttools==4.60.2
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.47.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
kiwisolver==1.4.7
Markdown==3.9
matplotlib==3.9.4
numpy==2.0.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pillow==11.3.0
proto-plus==1.27.0
protobuf==5.29.5
psycopg==3.2.13
psycopg-binary==3.2.13
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
reportlab==4.4.7
requests==2.32.5
rsa==4.9.1
six==1.17.0
soupsieve==2.8.1
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
zipp==3.23.0
